Support archiving brew components
Why:
We think every brew we produce is great, but unfortunately the market
doesn't always agree - before we go big and bottle 10000 units we push
out a small batch and test the market. During this trail phase we want
to save the brew components so we can reproduce it should it be a hit!

This change addresses the need by:
- Implementing NSSecureCoding on the CBCBrewComponents object.